Fixed issue of kernel still referring to a deleted version of Slicer
In the old implementation, slicer launcher path was set only once in kernel.json file and it was never changed.
This cause issue when Slicer was uninstalled and installed in another location, because the kernel.json file still referred to the old path.

Fixed this by always re-generating the kernel from a template when kernel is installed.