Always display install command
It is often easier to launch a command in a terminal than copy-pasting the Scripts folder path, so it makes sense to make this a bit more convenient to do.

Also, if a Python virtual environment is used then the virtual environment may need to be activated before jupyter-kernelspec can be launched. Since there is no standardized way of activating a virtual environment, it would be hard to make automatic installation work reliably.