Capture display content when display object is created
It was unexpected that just creating the display object did not actually do anything (e.g., did not change the view layout) until it was actually displayed.
Now we immediately capture the image when the display object is created and save them so that it can be displayed when _repr_mimebundle_ is called.