COMP: Fix python install to prevent main site-packages corruption
This commit reverts use of --force-reinstall originally introduced
in c59b08fa9 (Fix python package installation).

The use of --force-reinstall led to removal of package like
setuptools from the "site-packages" directory associated
with the application and was causing the build of following
extension (e.g SlicerRadiomics) to fail.

See https://discourse.slicer.org/t/factory-build-of-python-extensions-conflicting-with-each-other/21935

Co-authored-by: James Butler <jbutler@sonovol.com>